:sparkles: Add new command to generate board page for a specific resource
